+++ R/hotspot_gistar.R
@@ -27,7 +27,7 @@
 #' @param grid \code{\link[sf]{sf}} data frame containing polygons, which will
 #'   be used as the grid for which counts are made.
 #' @param weights \code{NULL} or the name of a column in \code{data} to be used
-#'   as weights for weighted counts and KDE values.
+#'   as weights for weighted counts, Gi*/Gi statistics and KDE values.
 #' @param nb_dist The distance around a cell that contains the neighbours of
 #'   that cell, which are used in calculating the statistic. If this argument is
 #'   \code{NULL} (the default), \code{nb_dist} is set as \code{cell_size *
@@ -85,7 +85,9 @@
 #' function to adjust the corresponding \eqn{p}-values for multiple comparison.
 #' The function also returns counts of points in each cell and (by default but
 #' optionally) kernel density estimates using the \code{\link[SpatialKDE]{kde}}
-#' function.
+#' function. If \code{weights} is supplied, the Gi*/Gi statistics are calculated
+#' from the weighted counts; otherwise, they are calculated from the unweighted
+#' counts.
 #'
 #' ## Coverage of the output data
 #'
@@ -246,7 +248,6 @@ hotspot_gistar <- function(
   # Calculate Gi*
   result <- gistar(
     counts,
-    n = "n",
     nb_dist = nb_dist,
     cell_size = cell_size,
     include_self = include_self,